Fix nits for rust 1.49.0 clippy
Some really cool new nits have been included, await over lock boundries
is going to really improve our lock based threading pattern for Rita.

+++ althea_kernel_interface/src/traffic_control.rs
@@ -238,7 +238,7 @@ impl dyn KernelInterface {
 
         // we need 1kbyte of burst cache per mbit of bandwidth to actually keep things
         // moving
-        let burst = bw * 1000 as u32;
+        let burst = bw * 1000u32;
         // amount of time a packet can spend in the burst cache, 40ms
         let latency = 40u32;
